[docs] Update end-of-file comment for Todo List Manager
Revise and expand the explanatory comment at the end of the file to provide a more comprehensive overview of the Todo List Manager application. The updated comment includes:

- Detailed descriptions of the main classes (Task and TodoList)
- Key features of the application
- User interface details
- Information about included unit tests
- Execution flow of the script
- A note on the educational purpose and lack of data persistence

This change improves the documentation and makes it easier for developers to understand the structure and functionality of the Todo List Manager application.