[feature] Add programmer joke to end of file
Add a programming-related joke about arrays at the end of the file.

Notes: The addition of this joke appears to be a lighthearted touch to the existing comments about potential future enhancements for the Todo List Manager. It doesn't affect the functionality of the code but adds a bit of humor to the file.