[feature] Add new programmer joke about dark mode
Add a new joke to the end of the file, playing on the double meaning
of "bugs" in programming. This joke combines the concept of light
attracting insects with software bugs, providing a humorous take on
why programmers prefer dark mode.

Notes: From the chat history, the user requested an "extremely funny" joke to give developers a good laugh. The AI assistant responded by adding this dark mode joke, which is a classic piece of programmer humor that should resonate well with the target audience.